# Some small optimizations