--- a/src/lib.rs
+++ b/src/lib.rs
@@ -62,7 +62,6 @@ mod scanners;
 mod entities;
 mod escape;
 mod puncttable;
-mod utils;
 mod parse;
 mod tree;
 mod linklabel;