I hope you aren't getting too tired of these optimization PRs. It appears there is still some juice left to squeeze in this thing. Getting rid of the `#[inline(always)]` and implementing the two event iterators separately yielded the largest speedup. Then there's a short circuit for html scanning and some other small scanner optimizations.

Master:
```
crdt_total              time:   [384.45 us 384.82 us 385.25 us]
```
PR:
```
crdt_total              time:   [356.17 us 357.33 us 358.61 us]
```
In total, this produces a ~7.5% speedup.