@marcusklaas 2019-04-27
```
crdt_total              time:   [350.39 us 350.70 us 351.05 us]
```