@raphlinus 2019-04-27
Not tired of the improvements at all! But am dealing with overload atm, might not be able to get to these in detail very soon.