@marcusklaas 2019-04-27
Yes!! Indeed, `as_bytes` is a no op! But it's the string subslicing that's expensive apparently!