@marcusklaas 2019-04-27
All the char boundary checking.. I'm not quite done yet, but you should give this PR a spin. It *flies* dude!