@marcusklaas 2019-04-27
Current benchmark time on my machine:
```
crdt_total              time:   [338.86 us 339.12 us 339.41 us]
```
or about 13.4% higher throughput!