@raphlinus 2019-04-27
On mine (13" MBP 2017), with target-cpu=native and simd feature:

```
crdt_total              time:   [258.98 us 259.69 us 260.59 us]
```

![Neo whoa gif]