@marcusklaas 2019-04-29
Thanks for the review!

In total these changes amount to a ~18% throughput increase on my machine.